Add 2.5m attachment node to Nom-O-Matic 25000
The Nom-O-Matic 25000 is a 2.5m part that tapers to a 1.25m attachment
point, which makes it awkward to connect to another 2.5m part.  This
change adds a 2.5m attachment node behind the 1.25m one, so it can be
connected properly to parts of either diameter.  (This is similar to the
multiple attachment nodes on the MKS Tundra modules, though without the
fancy conditional visibility on the tapered part.)