Fix Localization bug
There is a bug if playing KSP localized, where the save file gets corrupted when a Kerbal who was changed to Tourist after depleting supplies gets recovered and his trait should change back to his old trait.

But because USI-LS saves `experienceTrait.Title` (which is localized) and not `ExperienceTrait.TypeName` (which shouldn't be localized. But I'm not  entirely sure, this needs at least testing!), it will set the localized trait title which corrupts the save file.

Fixes #259